@@ -1,7 +1,7 @@
 name: path_provider_windows
 description: Windows implementation of the path_provider plugin
 homepage: https://github.com/flutter/plugins/tree/master/packages/path_provider/path_provider_windows
-version: 0.0.4+4
+version: 0.1.0-nullsafety
 
 flutter:
   plugin:
@@ -11,19 +11,19 @@ flutter:
         pluginClass: none
 
 dependencies:
-  path_provider_platform_interface: ^1.0.3
-  meta: ^1.0.5
-  path: ^1.6.4
+  path_provider_platform_interface: ^2.0.0-nullsafety
+  meta: ^1.3.0-nullsafety.6
+  path: ^1.8.0-nullsafety.3
   flutter:
     sdk: flutter
-  ffi: ^0.1.3
-  win32: ^1.7.1
+  ffi: ^0.2.0-nullsafety.1
+  win32: ^2.0.0-nullsafety.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0-nullsafety.3
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: ">=1.12.13+hotfix.4"
