@@ -13,7 +13,7 @@ class FakeVersionInfoQuerier implements VersionInfoQuerier {
 
   final Map<String, String> responses;
 
-  getStringValue(Pointer<Uint8> versionInfo, key) => responses[key];
+  getStringValue(Pointer<Uint8>? versionInfo, key) => responses[key];
 }
 
 void main() {
@@ -40,8 +40,11 @@ void main() {
       'ProductName': 'Amazing App',
     });
     final path = await pathProvider.getApplicationSupportPath();
-    expect(path, endsWith(r'AppData\Roaming\A Company\Amazing App'));
-    expect(Directory(path).existsSync(), isTrue);
+    expect(path, isNotNull);
+    if (path != null) {
+      expect(path, endsWith(r'AppData\Roaming\A Company\Amazing App'));
+      expect(Directory(path).existsSync(), isTrue);
+    }
   }, skip: !Platform.isWindows);
 
   test('getApplicationSupportPath with missing company', () async {
@@ -50,8 +53,11 @@ void main() {
       'ProductName': 'Amazing App',
     });
     final path = await pathProvider.getApplicationSupportPath();
-    expect(path, endsWith(r'AppData\Roaming\Amazing App'));
-    expect(Directory(path).existsSync(), isTrue);
+    expect(path, isNotNull);
+    if (path != null) {
+      expect(path, endsWith(r'AppData\Roaming\Amazing App'));
+      expect(Directory(path).existsSync(), isTrue);
+    }
   }, skip: !Platform.isWindows);
 
   test('getApplicationSupportPath with problematic values', () async {
@@ -61,12 +67,15 @@ void main() {
       'ProductName': r'A"/Terrible\|App?*Name',
     });
     final path = await pathProvider.getApplicationSupportPath();
-    expect(
-        path,
-        endsWith(r'AppData\Roaming\'
-            r'A _Bad_ Company_ Name\'
-            r'A__Terrible__App__Name'));
-    expect(Directory(path).existsSync(), isTrue);
+    expect(path, isNotNull);
+    if (path != null) {
+      expect(
+          path,
+          endsWith(r'AppData\Roaming\'
+              r'A _Bad_ Company_ Name\'
+              r'A__Terrible__App__Name'));
+      expect(Directory(path).existsSync(), isTrue);
+    }
   }, skip: !Platform.isWindows);
 
   test('getApplicationSupportPath with a completely invalid company', () async {
@@ -76,8 +85,11 @@ void main() {
       'ProductName': r'Amazing App',
     });
     final path = await pathProvider.getApplicationSupportPath();
-    expect(path, endsWith(r'AppData\Roaming\Amazing App'));
-    expect(Directory(path).existsSync(), isTrue);
+    expect(path, isNotNull);
+    if (path != null) {
+      expect(path, endsWith(r'AppData\Roaming\Amazing App'));
+      expect(Directory(path).existsSync(), isTrue);
+    }
   }, skip: !Platform.isWindows);
 
   test('getApplicationSupportPath with very long app name', () async {
